--- pmd-java/src/main/java/net/sourceforge/pmd/lang/java/oom/metrics/cyclo/CycloVisitor.java
@@ -1,12 +1,0 @@
-/**
- * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
- */
-
-package net.sourceforge.pmd.lang.java.oom.metrics.cyclo;
-
-import net.sourceforge.pmd.lang.java.ast.JavaParserVisitor;
-
-/** Marker interface for cyclo visitors. */
-public interface CycloVisitor extends JavaParserVisitor {
-
-}
