+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/oom/metrics/visitors/StandardCycloVisitor.java
@@ -2,7 +2,7 @@
  * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
  */
 
-package net.sourceforge.pmd.lang.java.oom.metrics.cyclo;
+package net.sourceforge.pmd.lang.java.oom.metrics.visitors;
 
 import org.apache.commons.lang3.mutable.MutableInt;
 
@@ -15,6 +15,7 @@ import net.sourceforge.pmd.lang.java.ast.ASTSwitchLabel;
 import net.sourceforge.pmd.lang.java.ast.ASTSwitchStatement;
 import net.sourceforge.pmd.lang.java.ast.ASTWhileStatement;
 import net.sourceforge.pmd.lang.java.ast.JavaNode;
+import net.sourceforge.pmd.lang.java.oom.metrics.visitors.CycloPathUnawareOperationVisitor;
 import net.sourceforge.pmd.lang.java.rule.codesize.NPathComplexityRule;
 
 /**
